GTDM_Lowlight/requirements.txt
absl-py==2.1.0
Brotli==1.0.9
certifi==2024.2.2
charset-normalizer==2.0.4
contourpy==1.2.1
cycler==0.12.1
einops==0.7.0
fonttools==4.50.0
grpcio==1.62.1
h5py==3.10.0
idna==3.4
intel-openmp==2024.1.0
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
mkl==2024.1.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
pip==23.3.1
protobuf==5.26.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.31.0
scipy==1.13.0
setuptools==68.2.2
six==1.16.0
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
torchsummary==1.5.1
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.2
wheel==0.41.2